update adapter project organization and delete tool
