+++ source/Sailfish.TestAdapter/TestProperties/SailfishTestTypeDefinition.cs
@@ -1,17 +1,6 @@
-using System;
 using Microsoft.VisualStudio.TestPlatform.ObjectModel;
 
 namespace Sailfish.TestAdapter.TestProperties;
-
-public class SailfishTestTypeDefinition
-{
-    internal static readonly TestProperty SailfishTestTypeDefinitionProperty = TestProperty.Register(
-        "Sailfish.TestTypeDefinition",
-        "TestTypeDefinition",
-        typeof(Type),
-        TestPropertyAttributes.None,
-        typeof(TestCase));
-}
 
 public class SailfishDisplayNameDefinition
 {
