--- source/Sailfish.TestAdapter/Utils/AttributeDiscoveryExtensionMethods.cs
@@ -1,54 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-
-namespace Sailfish.TestAdapter.Utils;
-
-internal static class AttributeDiscoveryExtensionMethods
-{
-    internal static bool HasAttribute<TAttribute>(this Type type, bool inherit = false) where TAttribute : Attribute
-    {
-        return type.GetCustomAttributes(typeof(TAttribute), inherit).Length > 0;
-    }
-
-    internal static bool HasAttribute<TAttribute>(this MethodInfo method, bool inherit = false) where TAttribute : Attribute
-    {
-        return method.GetCustomAttributes(typeof(TAttribute), inherit).Length > 0;
-    }
-
-    internal static bool HasAttribute<TAttribute>(this PropertyInfo property, bool inherit = false) where TAttribute : Attribute
-    {
-        return property.GetCustomAttributes(typeof(TAttribute), inherit).Length > 0;
-    }
-
-    internal static bool HasMethodWithAttribute<TAttribute>(this MethodInfo method) where TAttribute : Attribute
-    {
-        return method.GetCustomAttributes(typeof(TAttribute), false).Length > 0;
-    }
-
-    internal static IEnumerable<PropertyInfo> GetPropertiesWithAttribute<TAttribute>(this Type type) where TAttribute : Attribute
-    {
-        return type.GetProperties().Where(x => x.HasAttribute<TAttribute>());
-    }
-
-    internal static IEnumerable<MethodInfo> GetMethodsWithAttribute<TAttribute>(this Type type) where TAttribute : Attribute
-    {
-        return type.GetMethods().Where(x => x.HasAttribute<TAttribute>());
-    }
-
-    internal static IEnumerable<MethodInfo> GetMethodsWithAttribute<TAttribute>(this object instance) where TAttribute : Attribute
-    {
-        return instance.GetType().GetMethodsWithAttribute<TAttribute>();
-    }
-
-    internal static MethodInfo? GetMethodWithAttribute<TAttribute>(this object instance) where TAttribute : Attribute
-    {
-        return instance.GetType().GetMethodsWithAttribute<TAttribute>().SingleOrDefault();
-    }
-
-    internal static MethodInfo? GetMethodWithAttribute<TAttribute>(this Type type) where TAttribute : Attribute
-    {
-        return type.GetMethodsWithAttribute<TAttribute>().SingleOrDefault();
-    }
-}
