--- source/Sailfish.Tool/Sailfish.Tool.csproj
@@ -1,26 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <LangVersion>10</LangVersion>
-
-        <PackAsTool>true</PackAsTool>
-        <ToolCommandName>perf</ToolCommandName>
-        <PackageOutputPath>./nupkg</PackageOutputPath>
-        <RootNamespace>Sailfish.Tool</RootNamespace>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Autofac" Version="6.5.0" />
-        <PackageReference Include="Autofac.Extensions.DependencyInjection" Version="8.0.0" />
-        <PackageReference Include="McMaster.Extensions.CommandLineUtils" Version="4.0.2" />
-        <PackageReference Include="Serilog" Version="2.12.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <ProjectReference Include="..\Sailfish\Sailfish.csproj" />
-    </ItemGroup>
-
-</Project>
