+++ source/Sailfish.sln
@@ -6,8 +6,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Demo.API", "Demo.API\Demo.A
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests.Sailfish", "Tests.Sailfish\Tests.Sailfish.csproj", "{A42A41D3-7A3F-4475-B245-A2E896A2589F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sailfish.Tool", "Sailfish.Tool\Sailfish.Tool.csproj", "{B0C277A0-D2EC-4ABF-B342-C7F46A30EA2E}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AsAConsoleApp", "AsAConsoleApp\AsAConsoleApp.csproj", "{EE33D17C-3F19-40A9-8394-D966232723A6}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sailfish.TestAdapter", "Sailfish.TestAdapter\Sailfish.TestAdapter.csproj", "{71C488BD-3668-4D53-A5BD-6E9307215B5E}"
@@ -38,10 +36,6 @@ Global
 		{A42A41D3-7A3F-4475-B245-A2E896A2589F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A42A41D3-7A3F-4475-B245-A2E896A2589F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A42A41D3-7A3F-4475-B245-A2E896A2589F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B0C277A0-D2EC-4ABF-B342-C7F46A30EA2E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B0C277A0-D2EC-4ABF-B342-C7F46A30EA2E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B0C277A0-D2EC-4ABF-B342-C7F46A30EA2E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B0C277A0-D2EC-4ABF-B342-C7F46A30EA2E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EE33D17C-3F19-40A9-8394-D966232723A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EE33D17C-3F19-40A9-8394-D966232723A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE33D17C-3F19-40A9-8394-D966232723A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
